feat: add aliases for pulling test data
This adds aliases for pulling test data as well as helper
aliases that connect and disconnect from the VPN.

`pulltestdata` will pull the data from the test viper,
port 60000.

`pullviperdata` takes an argument of the port you want to pull
from.

The data scripts both use file paths that are on the connected
storage. This prevents the data from taking up space on the local
drive, but does restrict things to working only when the
additional drive is plugged in.